mission1_20251003/requirements.txt
beautifulsoup4==4.12.3
selenium==4.21.0
openai==1.30.5
tqdm
gradio==4.32.0
jupyter==1.0.0
jupyterlab==4.0.9
ipykernel==6.29.0
pandas==2.1.4
matplotlib==3.8.2
seaborn==0.13.0
python-dotenv==1.1.1
